HV: * changed all statistics-counting counter stuff to use 'counter_type' or       'ucounter_type' (consistency)     * added files counter.{h,cc} with, depending on compiled-in debuglevel,       typedef of ordinary volatile (u)int64_t or 'struct real_counter_type'       if DEBUG>=2.       If DEBUG>=2 then NO statistics counters will be kept. In fact, all       reads/writes to statistic counters disapper; the struct       real_counter_type will swallow everything and perform NO memory       access.       Rationale: when running tool like helgrind/valgrind, they complain       (completely correct) about unlocked/unsychronized reads+writes to       the statistics counters. So for debugging race conditions this       spams the output with correct but unuseful information. DEBUG>=2       will make /those/ messages disappear

+++ evlbi5a/mk5command/disk2file.cc
@@ -170,7 +170,7 @@ string disk2file_fn(bool qry, const vector<string>& args, runtime& rte ) {
            
             // Now it's safe to use 'd2f.*' 
             uint64_t start   = d2f.disk_args.pp_start.Addr;
-            uint64_t current = rte.statistics.counter(d2f.disk_stepid) + start;
+            uint64_t current = (uint64_t)rte.statistics.counter(d2f.disk_stepid) + start;
             uint64_t end     = d2f.disk_args.pp_end.Addr;
 
             reply << " 0 : active : " << d2f.file_name << " : " << start << " : " << current << " : "
